requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
annotated-types=0.7.0=pypi_0
anyio=4.4.0=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=2.4.1=pypi_0
astunparse=1.6.3=pypi_0
async-lru=2.0.4=pypi_0
attrs=23.2.0=pypi_0
babel=2.15.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
binaryornot=0.4.4=pypi_0
bleach=6.1.0=pypi_0
ca-certificates=2024.3.11=haa95532_0
certifi=2024.6.2=pypi_0
cffi=1.16.0=pypi_0
chardet=5.2.0=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
comm=0.2.2=pypi_0
contourpy=1.2.1=pypi_0
cookiecutter=2.6.0=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.8.1=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
dnspython=2.6.1=pypi_0
easydict=1.13=pypi_0
email-validator=2.2.0=pypi_0
exceptiongroup=1.2.1=pypi_0
executing=2.0.1=pypi_0
fastapi=0.111.1=pypi_0
fastapi-cli=0.0.4=pypi_0
fastjsonschema=2.20.0=pypi_0
flet=0.23.2=pypi_0
flet-core=0.23.2=pypi_0
flet-runtime=0.23.2=pypi_0
fonttools=4.53.0=pypi_0
fqdn=1.5.1=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.5=pypi_0
httptools=0.6.1=pypi_0
httpx=0.27.0=pypi_0
idna=3.7=pypi_0
importlib-metadata=7.2.1=pypi_0
importlib-resources=6.4.0=pypi_0
ipykernel=6.29.4=pypi_0
ipython=8.18.1=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
json5=0.9.25=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.22.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter-client=8.6.2=pypi_0
jupyter-core=5.7.2=pypi_0
jupyter-events=0.10.0=pypi_0
jupyter-lsp=2.2.5=pypi_0
jupyter-server=2.14.1=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyterlab=4.2.2=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.27.2=pypi_0
kiwisolver=1.4.5=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.9.0=pypi_0
matplotlib-inline=0.1.7=pypi_0
mdurl=0.1.2=pypi_0
mindspore=2.2.14=pypi_0
mistune=3.0.2=pypi_0
nbclient=0.10.0=pypi_0
nbconvert=7.16.4=pypi_0
nbformat=5.10.4=pypi_0
nest-asyncio=1.6.0=pypi_0
notebook-shim=0.2.4=pypi_0
numpy=2.0.0=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.10.0.84=pypi_0
openssl=3.0.14=h827c3e9_0
overrides=7.7.0=pypi_0
packaging=23.2=pypi_0
pandas=2.2.2=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pypi_0
pillow=10.3.0=pypi_0
pip=24.0=py39haa95532_0
platformdirs=4.2.2=pypi_0
prometheus-client=0.20.0=pypi_0
prompt-toolkit=3.0.47=pypi_0
protobuf=5.27.1=pypi_0
psutil=6.0.0=pypi_0
pure-eval=0.2.2=pypi_0
pyaml=24.4.0=pypi_0
pycparser=2.22=pypi_0
pydantic=2.8.2=pypi_0
pydantic-core=2.20.1=pypi_0
pygments=2.18.0=pypi_0
pyparsing=3.1.2=pypi_0
pypng=0.20220715.0=pypi_0
python=3.9.19=h1aa4202_1
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python-json-logger=2.0.7=pypi_0
python-multipart=0.0.9=pypi_0
python-slugify=8.0.4=pypi_0
pytz=2024.1=pypi_0
pywin32=306=pypi_0
pywinpty=2.0.13=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=26.0.3=pypi_0
qrcode=7.4.2=pypi_0
referencing=0.35.1=pypi_0
repath=0.9.0=pypi_0
requests=2.32.3=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rich=13.7.1=pypi_0
rpds-py=0.18.1=pypi_0
scikit-learn=1.5.0=pypi_0
scikit-optimize=0.10.2=pypi_0
scipy=1.13.1=pypi_0
seaborn=0.13.2=pypi_0
send2trash=1.8.3=pypi_0
setuptools=69.5.1=py39haa95532_0
shellingham=1.5.4=pypi_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.45.3=h2bbff1b_0
stack-data=0.6.3=pypi_0
starlette=0.37.2=pypi_0
terminado=0.18.1=pypi_0
text-unidecode=1.3=pypi_0
threadpoolctl=3.5.0=pypi_0
tinycss2=1.3.0=pypi_0
tomli=2.0.1=pypi_0
tornado=6.4.1=pypi_0
traitlets=5.14.3=pypi_0
typer=0.12.3=pypi_0
types-python-dateutil=2.9.0.20240316=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.1=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.2.2=pypi_0
uvicorn=0.30.3=pypi_0
vc=14.2=h2eaa2aa_1
vs2015_runtime=14.29.30133=h43f2093_3
watchdog=4.0.1=pypi_0
watchfiles=0.22.0=pypi_0
wcwidth=0.2.13=pypi_0
webcolors=24.6.0=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=pypi_0
websockets=12.0=pypi_0
wheel=0.43.0=py39haa95532_0
zipp=3.19.2=pypi_0
